+++ src/number.rs
@@ -2,30 +2,30 @@ use num::rational::Ratio;
 use std::fmt::Display;
 
 pub trait Number: Copy + Display + Eq {
-    fn from_int(x: i128) -> Self;
-    fn to_int(self) -> Option<i128>;
+    fn from_int(x: i64) -> Self;
+    fn to_int(self) -> Option<i64>;
 }
 
-impl Number for i128 {
+impl Number for i64 {
     #[inline]
-    fn from_int(x: i128) -> Self {
+    fn from_int(x: i64) -> Self {
         x
     }
 
     #[inline]
-    fn to_int(self) -> Option<i128> {
+    fn to_int(self) -> Option<i64> {
         Some(self)
     }
 }
 
-impl Number for Ratio<i128> {
+impl Number for Ratio<i64> {
     #[inline]
-    fn from_int(x: i128) -> Self {
-        Ratio::<i128>::from_integer(x)
+    fn from_int(x: i64) -> Self {
+        Ratio::<i64>::from_integer(x)
     }
 
     #[inline]
-    fn to_int(self) -> Option<i128> {
+    fn to_int(self) -> Option<i64> {
         if self.is_integer() {
             Some(*self.numer())
         } else {
